// frontend/web/src/services/API/proserve-wb-provisioning-api/models/MaintenanceWindow.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Product Provisioning API
 * API in charge of provisioning VEW products
 *
 * The version of the OpenAPI document: 2024-08-12T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MaintenanceWindow
 */
export interface MaintenanceWindow {
    /**
     * Day of the week for maintenance window.(monday-sunday)
     * @type {string}
     * @memberof MaintenanceWindow
     */
    day: string;
    /**
     * Start Time for maintenance window
     * @type {string}
     * @memberof MaintenanceWindow
     */
    startTime: string;
    /**
     * End Time for maintenance window
     * @type {string}
     * @memberof MaintenanceWindow
     */
    endTime: string;
}

/**
 * Check if a given object implements the MaintenanceWindow interface.
 */
export function instanceOfMaintenanceWindow(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "day" in value;
    isInstance = isInstance && "startTime" in value;
    isInstance = isInstance && "endTime" in value;

    return isInstance;
}

export function MaintenanceWindowFromJSON(json: any): MaintenanceWindow {
    return MaintenanceWindowFromJSONTyped(json, false);
}

export function MaintenanceWindowFromJSONTyped(json: any, ignoreDiscriminator: boolean): MaintenanceWindow {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'day': json['day'],
        'startTime': json['startTime'],
        'endTime': json['endTime'],
    };
}

export function MaintenanceWindowToJSON(value?: MaintenanceWindow | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'day': value.day,
        'startTime': value.startTime,
        'endTime': value.endTime,
    };
}